Move disposable-object.ts to /common
