+++ extensions/ql-vscode/src/common/discovery.ts
@@ -1,4 +1,4 @@
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "./disposable-object";
 import { getErrorMessage } from "../pure/helpers-pure";
 import { Logger } from "./logging";
