+++ extensions/ql-vscode/src/common/disposable-object.ts
