+++ extensions/ql-vscode/src/common/vscode/vscode-app.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import { VSCodeCredentials } from "./authentication";
-import { Disposable } from "../../pure/disposable-object";
+import { Disposable } from "../disposable-object";
 import { App, AppMode, EnvironmentContext } from "../app";
 import { AppEventEmitter } from "../events";
 import { extLogger, NotificationLogger, queryServerLogger } from "../logging";
