+++ extensions/ql-vscode/src/debugger/debugger-factory.ts
@@ -10,7 +10,7 @@ import {
 } from "vscode";
 import { isCanary } from "../config";
 import { LocalQueries } from "../local-queries";
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "../common/disposable-object";
 import { QueryRunner } from "../query-server";
 import { QLDebugConfigurationProvider } from "./debug-configuration";
 import { QLDebugSession } from "./debug-session";
