+++ extensions/ql-vscode/src/log-insights/summary-language-support.ts
@@ -11,7 +11,7 @@ import {
   window,
   workspace,
 } from "vscode";
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "../common/disposable-object";
 import { extLogger } from "../common";
 import { getErrorMessage } from "../pure/helpers-pure";
 import { SummaryLanguageSupportCommands } from "../common/commands";
