+++ extensions/ql-vscode/src/query-testing/test-adapter.ts
@@ -15,7 +15,7 @@ import {
 import { TestAdapterRegistrar } from "vscode-test-adapter-util";
 import { QLTestDiscovery } from "./qltest-discovery";
 import { Event, EventEmitter, CancellationTokenSource } from "vscode";
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "../common/disposable-object";
 import { CodeQLCliServer, TestCompleted } from "../codeql-cli/cli";
 import { testLogger } from "../common";
 import { TestRunner } from "./test-runner";
