+++ extensions/ql-vscode/src/query-testing/test-manager-base.ts
@@ -1,6 +1,6 @@
 import { copy, createFile, lstat, pathExists } from "fs-extra";
 import { TestUICommands } from "../common/commands";
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "../common/disposable-object";
 import { getActualFile, getExpectedFile } from "./test-adapter";
 import { TestItem, TextDocumentShowOptions, Uri, window } from "vscode";
 import { basename } from "path";
