+++ extensions/ql-vscode/src/query-testing/test-ui.ts
@@ -8,7 +8,7 @@ import {
   TestSuiteEvent,
 } from "vscode-test-adapter-api";
 import { TestTreeNode } from "./test-tree-node";
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "../common/disposable-object";
 import { QLTestAdapter } from "./test-adapter";
 import { App } from "../common/app";
 import { TestManagerBase } from "./test-manager-base";
