+++ extensions/ql-vscode/src/variant-analysis/gh-api/mocks/mock-gh-api-server.ts
@@ -2,7 +2,7 @@ import { join, resolve } from "path";
 import { pathExists } from "fs-extra";
 import { setupServer, SetupServer } from "msw/node";
 
-import { DisposableObject } from "../../../pure/disposable-object";
+import { DisposableObject } from "../../../common/disposable-object";
 
 import { Recorder } from "./recorder";
 import { createRequestHandlers } from "./request-handlers";
