+++ extensions/ql-vscode/src/variant-analysis/variant-analysis-monitor.ts
@@ -10,7 +10,7 @@ import {
 } from "./shared/variant-analysis";
 import { VariantAnalysis as ApiVariantAnalysis } from "./gh-api/variant-analysis";
 import { processUpdatedVariantAnalysis } from "./variant-analysis-processor";
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "../common/disposable-object";
 import { sleep } from "../pure/time";
 import { getErrorMessage } from "../pure/helpers-pure";
 import { App } from "../common/app";
