+++ extensions/ql-vscode/test/unit-tests/pure/disposable-object.test.ts
@@ -1,4 +1,4 @@
-import { DisposableObject } from "../../../src/pure/disposable-object";
+import { DisposableObject } from "../../../src/common/disposable-object";
 
 describe("DisposableObject and DisposeHandler", () => {
   const disposable1 = {
