+++ extensions/ql-vscode/test/vscode-tests/cli-integration/debugger/debug-controller.ts
@@ -8,7 +8,7 @@ import {
   workspace,
 } from "vscode";
 import * as CodeQLProtocol from "../../../../src/debugger/debug-protocol";
-import { DisposableObject } from "../../../../src/pure/disposable-object";
+import { DisposableObject } from "../../../../src/common/disposable-object";
 import { QueryResultType } from "../../../../src/pure/legacy-messages";
 import { CoreCompletedQuery } from "../../../../src/query-server/query-runner";
 import { QueryOutputDir } from "../../../../src/run-queries-shared";
